requirements.txt
huggingface-hub==0.25.2
torch
torchvision
torchaudio
diffusers
transformers
safetensors
moviepy
tortoise-tts
pyyaml
